test(GMPExecutable): moving into separate files
